% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dgp}
\alias{dgp}
\title{Data Generating Process for Spatial Scalar-on-Function Model}
\usage{
dgp(n, rho, out.p = 0)
}
\arguments{
\item{n}{Integer. Sample size.}

\item{rho}{Numeric. Spatial autoregressive parameter.}

\item{out.p}{Numeric. Proportion of outliers to introduce (default is 0).}
}
\value{
A list with the following elements:
\describe{
\item{y}{Response vector of length \code{n}.}
\item{x}{Matrix of functional predictor values.}
\item{w}{Spatial weight matrix of size \code{n x n}.}
\item{tcoefs}{True coefficient function used to generate the data.}
\item{out.index}{Indices of outlier observations, if any.}
}
}
\description{
Generates synthetic data from a spatial scalar-on-function regression model.
}
\examples{
\dontrun{
sim_data <- dgp(n = 250, rho=0.5)
y <- sim_data$y
x <- sim_data$x
w <- sim_data$w
tcoef <- sim_data$tcoefs
}

}
