% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCA.R
\name{getPCA}
\alias{getPCA}
\title{Functional Principal Component Analysis with B-spline Basis}
\usage{
getPCA(data, nbasis, gp)
}
\arguments{
\item{data}{Numeric matrix of size \code{n x p}. Functional data observed over grid points in rows.}

\item{nbasis}{Integer. Number of B-spline basis functions.}

\item{gp}{Numeric vector of length \code{p}. Grid points corresponding to functional observations.}
}
\value{
A list with the following components:
\describe{
\item{PCAcoef}{An \code{fd} object representing estimated eigenfunctions.}
\item{PCAscore}{Matrix of principal component scores.}
\item{meanScore}{\code{fd} object for the mean function.}
\item{ncomp}{Number of components explaining more than 95\\% of the variance.}
\item{bs_basis}{B-spline basis used.}
\item{evalbase}{Evaluation of basis functions at grid points.}
\item{gp}{The original grid points used.}
}
}
\description{
Performs FPCA (Functional Principal Component Analysis) using a B-spline basis.
}
