% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_wei.R
\name{norm_wei}
\alias{norm_wei}
\title{Row-Normalize a Spatial Weight Matrix}
\usage{
norm_wei(W, tol = 1e-08)
}
\arguments{
\item{W}{Numeric matrix. Spatial weight matrix of size \code{n x n}.}

\item{tol}{Numeric. Tolerance for checking whether rows sum to 1. Default is \code{1e-8}.}
}
\value{
A row-normalized version of the input matrix \code{W}.
}
\description{
Checks whether a spatial weight matrix is row-normalized. If not, it normalizes the matrix by dividing each row by its sum.
}
