% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_fqrm.R
\name{predict_fqrm}
\alias{predict_fqrm}
\title{Predict Method for Functional Quantile Regression with Spatial Effects}
\usage{
predict_fqrm(object, xnew, wnew)
}
\arguments{
\item{object}{A fitted model object returned by \code{\link{fqrm}}.}

\item{xnew}{Numeric matrix of size \code{n x p}. New functional predictor data (on the same grid used in training).}

\item{wnew}{Numeric matrix of size \code{n x n}. Spatial weight matrix for the new observations.}
}
\value{
A numeric vector of predicted values of length \code{n}.
}
\description{
Makes predictions for new observations using a fitted spatial functional quantile regression model.
}
\examples{
\dontrun{
sim_data <- dgp(n = 250, rho=0.5)
y <- sim_data$y
x <- sim_data$x
w <- sim_data$w
fit_kim <- fqrm(y=y, x=x, w=w, tau=tau, method = "KM")
fit_ch <- fqrm(y=y, x=x, w=w, tau=tau, method = "Ch")
sim_test <- dgp(n = 1000, rho=rho)
y_test <- sim_test$y
x_test <- sim_test$x
w_test <- sim_test$w
predict_kim <- predict_fqrm(object = fit_kim, xnew = x_test, wnew = w_test)
predict_ch <- predict_fqrm(object = fit_ch, xnew = x_test, wnew = w_test)
}

}
