% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssofqrm-package.R
\docType{package}
\name{ssofqrm-package}
\alias{ssofqrm}
\alias{ssofqrm-package}
\title{Spatial Scalar-on-Function Quantile Regression Model}
\description{
Functions for implementing spatial scalar-on-function quantile regression model.
}
\seealso{
\code{\link{fqrm}}, \code{\link{dgp}}, \code{\link{getPCA}}, etc.
}
\author{
Muge Mutis (\email{muge.mutis@yildiz.edu.tr}) \link{aut, cre, cph} \cr
Ufuk Beyaztas \link{aut} \cr
Filiz Karaman \link{aut} \cr
Han Lin Shang \link{aut}
}
